98_exercises/hashing/pyproject.toml
[tool.poetry]
name = "hashing"
version = "0.1.0"
description = ""
authors = ["mija0001 <mija0001@stud.kea.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
argon2-cffi = "^21.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
